[includes: review dependencies]
Ensure all files only include what they actually need.
Sort includes.

+++ action/layout.c
@@ -1,14 +1,13 @@
-#include <config.h>
 #include <core/dwm.h>
-#include <actions.h>
-#include <xlib/window.h>
 #include <core/statusbar.h>
+#include <actions.h>
+#include <config.h>
 
 
 /* global functions */
 void action_layout_select(action_arg_t const *arg){
 	if(arg == 0x0 || arg->v == 0x0){
-		for(unsigned int i=0; i<nlayouts; i++){
+		for(size_t i=0; i<nlayouts; i++){
 			if(dwm.layout == layouts + i){
 				i = (i + 1 < nlayouts) ? i + 1 : 0;
 				dwm.layout = layouts + i;
