+++ config/tags.c
@@ -1,8 +1,9 @@
-#include <utils/math.h>
+#include <stddef.h>
+#include <utils/utils.h>
 
 
 /* global variables */
 char const *tags[] = { "💻", "💻", "🗇", "📟" };
-unsigned int ntags = LENGTH(tags);
+size_t ntags = LENGTH(tags);
 
 static_assert(LENGTH(tags) < 32, "too many tags, number of tags must fit into an unsigned int bit array");
